+++ src/main/java/com/program/service/TeacherService.java
@@ -1,9 +1,7 @@
 package com.program.service;
 
 import com.program.exception.TeacherException;
-import com.program.exception.UserException;
-import com.program.model.Teacher;
-import com.program.model.User;
+import com.program.model.teacher.Teacher;
 
 import java.util.List;
 
@@ -11,6 +9,6 @@ public interface TeacherService {
 
     List<Teacher> findAllTeacher() throws TeacherException;
 
-    Teacher getTeacherById(Integer id) throws TeacherException;
+    Teacher getTeacherById(Long id) throws TeacherException;
 
 }
